Make sure we use the new decoder in the native clients

+++ bin/benchmark_nc.py
@@ -240,7 +240,7 @@ def delete_tree(dir):
     except IOError:
         print('No remote directory: %s' % dir)
 
-def setup_tempdir(dir, models, wav, alphabet, binaries):
+def setup_tempdir(dir, models, wav, alphabet, lm_binary, trie, binaries):
     r'''
     Copy models, libs and binary to a directory (new one if dir is None)
     '''
@@ -269,8 +269,8 @@ def setup_tempdir(dir, models, wav, alphabet, binaries):
                 print('Copying %s to %s' % (f, dir))
                 shutil.copy2(f, dir)
 
-    for extra_file in [ wav, alphabet ]:
-        if not os.path.isfile(os.path.join(dir, os.path.basename(extra_file))):
+    for extra_file in [ wav, alphabet, lm_binary, trie ]:
+        if extra_file and not os.path.isfile(os.path.join(dir, os.path.basename(extra_file))):
             print('Copying %s to %s' % (extra_file, dir))
             shutil.copy2(extra_file, dir)
 
@@ -376,7 +376,7 @@ def establish_ssh(target=None, auto_trust=False, allow_agent=True, look_keys=Tru
 
     return ssh_conn
 
-def run_benchmarks(dir, models, wav, alphabet, iters=-1, extra_aot_model=None):
+def run_benchmarks(dir, models, wav, alphabet, lm_binary=None, trie=None, iters=-1, extra_aot_model=None):
     r'''
     Core of the running of the benchmarks. We will run on all of models, against
     the WAV file provided as wav, and the provided alphabet.
@@ -401,7 +401,11 @@ def run_benchmarks(dir, models, wav, alphabet, iters=-1, extra_aot_model=None):
           'stddev': numpy.infty
         }
 
-        cmdline = './deepspeech "%s" "%s" "%s" -t' % (model_filename, wav, alphabet)
+        if lm_binary and trie:
+            cmdline = './deepspeech "%s" "%s" "%s" "%s" "%s" -t' % (model_filename, wav, alphabet, lm_binary, trie)
+        else:
+            cmdline = './deepspeech "%s" "%s" "%s" -t' % (model_filename, wav, alphabet)
+
         for it in range(iters):
             sys.stdout.write('\rRunning %s: %d/%d' % (os.path.basename(model), (it+1), iters))
             sys.stdout.flush()
@@ -459,6 +463,10 @@ def handle_args():
                                  help='WAV file to pass to native_client. Supply again in plotting mode to draw realine line.')
     parser.add_argument('--alphabet', type=str, required=False,
                                  help='Text file to pass to native_client for the alphabet.')
+    parser.add_argument('--lm_binary', type=str, required=False,
+                                 help='Path to the LM binary file used by the decoder.')
+    parser.add_argument('--trie', type=str, required=False,
+                                 help='Path to the trie file used by the decoder.')
     parser.add_argument('--iters', type=int, required=False, default=5,
                                  help='How many iterations to perfom on each model.')
     parser.add_argument('--keep', required=False, action='store_true',
@@ -497,13 +505,19 @@ def do_main():
     global ssh_conn
     ssh_conn = establish_ssh(target=cli_args.target, auto_trust=cli_args.autotrust, allow_agent=cli_args.allowagent, look_keys=cli_args.lookforkeys)
 
-    tempdir, sorted_models = setup_tempdir(dir=cli_args.dir, models=cli_args.models, wav=cli_args.wav, alphabet=cli_args.alphabet, binaries=cli_args.binaries)
+    tempdir, sorted_models = setup_tempdir(dir=cli_args.dir, models=cli_args.models, wav=cli_args.wav, alphabet=cli_args.alphabet, lm_binary=cli_args.lm_binary, trie=cli_args.trie, binaries=cli_args.binaries)
 
     dest_sorted_models = map(lambda x: os.path.join(tempdir, os.path.basename(x)), sorted_models)
     dest_wav = os.path.join(tempdir, os.path.basename(cli_args.wav))
     dest_alphabet = os.path.join(tempdir, os.path.basename(cli_args.alphabet))
 
-    inference_times = run_benchmarks(dir=tempdir, models=dest_sorted_models, extra_aot_model=cli_args.so_model, wav=dest_wav, alphabet=dest_alphabet, iters=cli_args.iters)
+    if cli_args.lm_binary and cli_args.trie:
+        dest_lm_binary = os.path.join(tempdir, os.path.basename(cli_args.lm_binary))
+        dest_trie = os.path.join(tempdir, os.path.basename(cli_args.trie))
+        inference_times = run_benchmarks(dir=tempdir, models=dest_sorted_models, extra_aot_model=cli_args.so_model, wav=dest_wav, alphabet=dest_alphabet, lm_binary=dest_lm_binary, trie=dest_trie, iters=cli_args.iters)
+    else:
+        inference_times = run_benchmarks(dir=tempdir, models=dest_sorted_models, extra_aot_model=cli_args.so_model, wav=dest_wav, alphabet=dest_alphabet, iters=cli_args.iters)
+
     if cli_args.csv:
         produce_csv(input=inference_times, output=cli_args.csv)
